+++ core/plugin/cubism3_layer/Cubism3Layer.js
@@ -1,7 +1,7 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.Cubism3Layer = void 0;
-const { Layer, argChk_Num } = require('skynovel/web');
+const { Layer, argChk_Num } = require('@famibee/skynovel/web');
 const pixi_js_1 = require("pixi.js");
 class Cubism3Layer extends Layer {
     constructor() {