+++ core/web4webpack.js
@@ -12,7 +12,7 @@ if (location.pathname.slice(-15) == '/index_app.html') {
 	pcur = '';
 	dip['oninit_run'] = false;
 }
-const {SysWeb} = require('skynovel/web');
+const {SysWeb} = require('@famibee/skynovel/web');
 const sys = new SysWeb(hPlg, {cur: pcur +'top/', dip: JSON.stringify(dip)});
 globalThis.runSN = (cur = pcur +'top')=> sys.run(cur);
 globalThis.stopSN = ()=> sys.stop();